# kogi-infra/openapi/kogi-server.yaml
openapi: 3.0.3
info:
  title: Kogi MVP Server API
  version: 0.2.0
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/engine/control:
    post:
      summary: Control engine service (start/stop/reload)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
      responses:
        "200":
          description: engine control accepted
  /api/v1/engine/ingest:
    post:
      summary: Ingest payload into engine service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: engine ingest accepted
...